// src/models/IEidCustomer.ts
export class IEidCustomer{
    lastName?: string;
    gender?: string;
    fullNameAR?: string;
    postalCode?: string;
    dateOfBirth?: string;
    fullNameEN?: string;
    arabicNationality?: string;
    expiryDate?: string;
    firstName?: string;
    phoneNumber?: string;
    nationality?: string;
    issueDate?: string;
    cardNumber?: string;
    photo?:string;
}